Development Continues on XMLtoDICT build
